# docs/swagger.yaml
basePath: /
definitions:
  entity.ChangePassword:
    properties:
      oldPassword:
        type: string
      password:
        type: string
      userId:
        type: string
      username:
        type: string
    type: object
  entity.Login:
    properties:
      password:
        type: string
      userid:
        type: string
    type: object
...
  service.User:
    properties:
      avatar:
        type: string
      email:
        type: string
      emailverified:
        type: integer
      password:
        type: string
      sex:
        type: integer
      userid:
        type: string
      username:
        type: string
    type: object
  utils.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: revolveyao@qq.com
    name: API支持者
    url: http://www.swagger.io/support
  description: 星辰大海后端API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 星辰大海 API
  version: "1.0"
paths:
...
  /v1/login:
    post:
      consumes:
      - application/x-www-form-urlencoded
      - application/json
      description: 登陆
      operationId: login-main-json
      parameters:
      - description: 用户登陆信息
        in: body
        name: User
        required: true
        schema:
          $ref: '#/definitions/entity.Login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: 登陆
...
  /v1/refresh_toekn:
    get:
      description: 刷新token
      operationId: login-refresh-toekn-json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: 刷新token
  /v1/register:
    post:
      consumes:
      - application/json
      description: 注册用户
      operationId: register-main-json
      parameters:
      - description: 用户注册信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/service.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: 注册
...
  /v1/user/info/{userid}:
    get:
      consumes:
      - application/json
      description: 根据用户id获取用户信息
      operationId: get-user-info-json
      parameters:
      - description: 用户名
        in: path
        name: userid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: 获取用户信息
  /v1/user/updateinfo:
    post:
      consumes:
      - application/json
      description: 根据传过来的值修改用户信息
      operationId: update-user-info-json
      parameters:
      - description: 用户注册信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/service.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: 修改用户信息
  /v1/user/updatepass:
    post:
      consumes:
      - application/json
      description: 根据传过来的值修改用户密码
      operationId: update-user-password-json
      parameters:
      - description: 用户密码修改
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/entity.ChangePassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: 修改用户密码
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"